Make room IDs deterministic and track nylo parents
This changes the way that room IDs are assigned to NPCs. Previously, the
room ID was simply the NPC object's hash code, which is entirely local
to a single client. This attempts to make room IDs more deterministic by
deriving them from information about the NPC as it spawns, allowing
separate clients to independently arrive at the same room ID in the
ideal case.

Additionally, this updates Nylocas tracking to attempt to assign a
parent to each split that spawns in the room, and transmit its room ID
as part of the update event.