Start raids in response to chat messages
The in-game raid start chat message is a natural point of
synchronization between Blert clients, as the game sends it to all
party members at once. This makes it a preferred way to start over
the raid varbit, which is staggered on player entry. Additionally,
completed party information is available through the party widget
instead of having to wait for orbs to populate.