Don't end ToB raids on varbit change
Logging out and back in while in a raid temporarily sets the ToB status
varbit to 0 on login. This caused the raid to end prematurely. This
updates the ToB manager to only end a raid once the player has left ToB
and returned to the lobby. Tracking this requires additional challenge
states to distinguish between having left a raid and being in the lobby
before entering.