Fix spawns for bosses which exist on entry
The data tracker code for bosses like Bloat, which are already spawned
when its room is entered, assumed that a spawn event would be dispatched
on initialization. This no longer seems to be the case. This change
updates these data trackers to scan for their boss on room start
instead.