Synchronize ToB boss hitpoints with varbit
There is a limit to how many hitsplats a RuneScape client receives from
the server. Therefore, tracking hitpoints solely from hitsplats can lead
to undercalculating the amount of damage an NPC has taken.

For ToB bosses, there is a varbit which indicates the boss HP to a tenth
of a percent. This updates the code to listen for updates of that
varbit, and check the calcluated hitpoints against its value, updating
the tracked hitpoints if they differ sufficiently.