Fix plugin to work with OSRS actor update
A recent OSRS update changed the behavior of cached actors: now, clients
only receive events for actors which are in render distance instead of
all actors in an instance.

This updates the plugin to not rely on NPC spawns for certain tracking
events, such as Maiden room start. Additionally, updates to remove use
of deprecated RuneLite APIs.