Track Xarpus exhumeds and poison splats
Defines events for Xarpus' P1 exhumeds and P2 poison splats, and updates
the Xarpus tracker to dispatch them.

Exhumed events are sent for each spanwed exhumed after it despawns, with
the ticks on which it healed the boss (if any).

Splat events are sent whenever a new poison splat lands, indicating the
source of the splat (Xarpus spit or a bounce from another splat).